posts-backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.4.4
tzdata==2024.1
pillow==10.3.0